Set warning level 4 and treat warnings as errors

+++ source/widgets/Window.h
@@ -12,7 +12,7 @@ public:
 	void Render(const RenderArea& targetArea) override;
 
 private:
-	int m_Width, m_Height;
+	float m_Width, m_Height;
 
 	virtual RenderArea GetAreaForChild(const RenderArea& availableArea, RenderArea& usedArea, RenderContext& ctx) const override;
 };
